+++ src/kong_backend/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kong_backend"
-version = "0.0.12"
+version = "0.0.13"
 edition = "2021"
 description = "Kong Swap backend canister"
