+++ src/kong_backend/src/lib.rs
@@ -34,4 +34,4 @@ mod user;
 mod user_balances;
 
 pub const APP_NAME: &str = "Kong Swap";
-pub const APP_VERSION: &str = "v0.0.12";
+pub const APP_VERSION: &str = "v0.0.13";
