+++ src/kong_backend/src/swap/mod.rs
@@ -12,3 +12,4 @@ pub mod swap_reply_helpers;
 pub mod swap_transfer;
 pub mod swap_transfer_from;
 pub mod update_liquidity_pool;
+pub mod archive_to_kong_data;
