+++ src/kong_faucet/src/lib.rs
@@ -5,4 +5,4 @@ mod stable_memory;
 mod stable_user;
 
 pub const APP_NAME: &str = "Kong Swap Faucet";
-pub const APP_VERSION: &str = "v0.0.12";
+pub const APP_VERSION: &str = "v0.0.13";
