+++ src/kong_lib/src/lib.rs
@@ -29,4 +29,4 @@ pub mod user;
 mod user_balances;
 
 pub const APP_NAME: &str = "Kong Lib";
-pub const APP_VERSION: &str = "v0.0.12";
+pub const APP_VERSION: &str = "v0.0.13";
